requirements.txt
requests==2.24.0
selenium==3.141.0
beautifulsoup4==4.9.3
click==7.1.2
matplotlib==3.3.2
pandas==1.1.4
progressbar33==2.4
python_dateutil==2.8.1
tqdm==4.51.0
